pyproject.toml
[project]
name = "neuro-forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitsandbytes>=0.49.1",
    "causal-conv1d>=1.6.0",
    "einops>=0.8.2",
    "hydra-core>=1.3.2",
    "lightning>=2.6.1",
    "mamba-ssm>=2.3.0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchao>=0.16.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "triton>=3.6.0",
    "wandb>=0.25.0",
]
